Fix JavaScript error caused by modifications for #1980.
Previously, I added code that dependes on jQuery UI
Autocomplete to the main admin.js script. However,
I added jquery-ui-autocomplete as a dependency of the
csv-import.js script, that is included only in the
CSV Import screens. As a result, the plugin started
triggering a JS error in every other admin page.

I moved all the code to handle the CSV Import form from
the admin.js to csv-import.js. Right now it makes more
sense to have it there, because that's the only place
where the autocomplete field is being used.