Fix Query Integration when CPT Compat Mode is on (#2067).
When CPT advanced integration is disabled, the Query Integration
class was unable to set the necessary flags and update query vars
with information such as posts_per_page, order and orderby (for
Category and Tags archives).

The reason is that when CPT Compat Mode is on, the rewrite rules
do not use the Taxonomies identifiers, instead they include the
configured category and tag slugs.

This patch updates Query Integration to check for the category
and tag slugs as indicators of being in a Category or Tag view,
respectively.

Also, it udpdates the CPT Compat Mode class to preserve posts_per_page,
order and orderby when it replaces the current query with a simpler?
one.