Fix fatal error with Russian (#361)
* Fix fatal error with Russian
This line was causing "Unknown format specifier" error:
$feed_title = sprintf( _x( '%s Feed', 'rss feed', 'business-directory-plugin' ), $main_page_title );
Changing 'rss feed' to 'rss' fixes the issue. It seems a combination of %s and a modifier name with a space is the problem, but not 100% sure. It might only happen with 'rss feed' as the name. Updating a few others just in case.